requirements.txt
# Core dependencies
python-dotenv==1.0.0

# Email parsing
# IMAP is built-in to Python

# Google Drive (for WhatsApp sync)
google-auth==2.23.4
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0

# Web tools
requests==2.31.0

# Encryption (optional - for WhatsApp DB decryption)
pycryptodome==3.19.0

# AI Agent (Google Gemini)
google-generativeai==0.8.3

# AI Agent (Minimax - L4 için)
anthropic==0.39.0  # Minimax Anthropic SDK ile kullanılıyor

# LLM (optional - for future use)
# openai==1.3.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
